refactor(pathPolicy): 이동경로 요금 관련 클래스 패키지 이동
