+++ src/main/java/nextstep/subway/fare/exception/PayZoneFaultException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.path.exception;
+package nextstep.subway.fare.exception;
 
 public class PayZoneFaultException extends RuntimeException {
     public PayZoneFaultException(String message) {
