--- src/main/java/nextstep/subway/path/exception/CalculationFailException.java
@@ -1,7 +1,0 @@
-package nextstep.subway.path.exception;
-
-public class CalculationFailException extends RuntimeException {
-    public CalculationFailException(String message) {
-        super(message);
-    }
-}
